+++ uni_modules/uni-collapse/components/uni-collapse/uni-collapse.vue
@@ -60,13 +60,14 @@
 			this.names = []
 		},
 		mounted() {
-			this.setOpen(this.dataValue)
+			this.$nextTick(()=>{
+				this.setOpen(this.dataValue)
+			})
 		},
 		methods: {
 			setOpen(val) {
 				let str = typeof val === 'string'
 				let arr = Array.isArray(val)
-
 				this.childrens.forEach((vm, index) => {
 					if (str) {
 						if (val === vm.nameSync) {
@@ -131,7 +132,7 @@
 		}
 	}
 </script>
-<style lang="scss" scoped>
+<style lang="scss" >
 	.uni-collapse {
 		/* #ifndef APP-NVUE */
 		width: 100%;