+++ uni_modules/uni-fab/components/uni-fab/uni-fab.vue
@@ -217,7 +217,7 @@
 	}
 </script>
 
-<style lang="scss" scoped>
+<style lang="scss" >
 	$uni-shadow-base:0 1px 5px 2px rgba($color: #000000, $alpha: 0.3) !default;
 
 	.uni-fab {