+++ uni_modules/uni-nav-bar/components/uni-nav-bar/uni-nav-bar.vue
@@ -1,52 +1,59 @@
 <template>
-	<view class="uni-navbar">
-		<view :class="{ 'uni-navbar--fixed': fixed, 'uni-navbar--shadow': shadow, 'uni-navbar--border': border }" :style="{ 'background-color': backgroundColor }"
-		 class="uni-navbar__content">
+	<view class="uni-navbar" :class="{'uni-dark':dark}">
+		<view :class="{ 'uni-navbar--fixed': fixed, 'uni-navbar--shadow': shadow, 'uni-navbar--border': border }"
+			:style="{ 'background-color': themeBgColor }" class="uni-navbar__content">
 			<status-bar v-if="statusBar" />
-			<view :style="{ color: color,backgroundColor: backgroundColor }" class="uni-navbar__header uni-navbar__content_view">
-				<view @tap="onClickLeft" class="uni-navbar__header-btns uni-navbar__header-btns-left uni-navbar__content_view">
-					<view class="uni-navbar__content_view" v-if="leftIcon.length">
-						<uni-icons :color="color" :type="leftIcon" size="22" color="#333"/>
-					</view>
-					<view :class="{ 'uni-navbar-btn-icon-left': !leftIcon.length }" class="uni-navbar-btn-text uni-navbar__content_view"
-					 v-if="leftText.length">
-						<text :style="{ color: color, fontSize: '12px' }">{{ leftText }}</text>
-					</view>
-					<slot name="left" />
+			<view :style="{ color: themeColor,backgroundColor: themeBgColor ,height:navbarHeight}"
+				class="uni-navbar__header">
+				<view @tap="onClickLeft" class="uni-navbar__header-btns uni-navbar__header-btns-left"
+					:style="{width:leftIconWidth}">
+					<slot name="left">
+						<view class="uni-navbar__content_view" v-if="leftIcon.length > 0">
+							<uni-icons :color="themeColor" :type="leftIcon" size="20" />
+						</view>
+						<view :class="{ 'uni-navbar-btn-icon-left': !leftIcon.length > 0 }" class="uni-navbar-btn-text"
+							v-if="leftText.length">
+							<text :style="{ color: themeColor, fontSize: '12px' }">{{ leftText }}</text>
+						</view>
+					</slot>
 				</view>
-				<view class="uni-navbar__header-container uni-navbar__content_view" @tap="onClickTitle">
-					<view class="uni-navbar__header-container-inner uni-navbar__content_view" v-if="title.length">
-						<text class="uni-nav-bar-text" :style="{color: color }">{{ title }}</text>
-					</view>
-					<!-- 标题插槽 -->
-					<slot />
+				<view class="uni-navbar__header-container " @tap="onClickTitle">
+					<slot>
+						<view class="uni-navbar__header-container-inner" v-if="title.length>0">
+							<text class="uni-nav-bar-text uni-ellipsis-1"
+								:style="{color: themeColor }">{{ title }}</text>
+						</view>
+					</slot>
 				</view>
-				<view :class="title.length ? 'uni-navbar__header-btns-right' : ''" @tap="onClickRight" class="uni-navbar__header-btns uni-navbar__content_view">
-					<view class="uni-navbar__content_view uni-navbar__header-btns-right" v-if="rightIcon.length">
-						<uni-icons :color="color" :type="rightIcon" size="22" />
-					</view>
-					<!-- 优先显示图标 -->
-					<view class="uni-navbar-btn-text uni-navbar__content_view" v-if="rightText.length && !rightIcon.length">
-						<text class="uni-nav-bar-right-text">{{ rightText }}</text>
-					</view>
-					<slot name="right" />
+				<view @click="onClickRight" class="uni-navbar__header-btns uni-navbar__header-btns-right"
+					:style="{width:rightIconWidth}">
+					<slot name="right">
+						<view v-if="rightIcon.length">
+							<uni-icons :color="themeColor" :type="rightIcon" size="22" />
+						</view>
+						<view class="uni-navbar-btn-text" v-if="rightText.length && !rightIcon.length">
+							<text class="uni-nav-bar-right-text" :style="{ color: themeColor}">{{ rightText }}</text>
+						</view>
+					</slot>
 				</view>
 			</view>
 		</view>
 		<view class="uni-navbar__placeholder" v-if="fixed">
 			<status-bar v-if="statusBar" />
-			<view class="uni-navbar__placeholder-view" />
+			<view class="uni-navbar__placeholder-view" :style="{ height:navbarHeight}" />
 		</view>
 	</view>
 </template>
 
 <script>
 	import statusBar from "./uni-status-bar.vue";
+	const getVal = (val) => typeof val === 'number' ? val + 'px' : val;
 
 	/**
 	 * NavBar 自定义导航栏
 	 * @description 导航栏组件，主要用于头部导航
 	 * @tutorial https://ext.dcloud.net.cn/plugin?id=52
+	 * @property {Boolean} dark 开启黑暗模式
 	 * @property {String} title 标题文字
 	 * @property {String} leftText 左侧按钮文本
 	 * @property {String} rightText 右侧按钮文本
@@ -66,8 +73,12 @@
 		components: {
 			statusBar
 		},
-		emits:['clickLeft','clickRight','clickTitle'],
+		emits: ['clickLeft', 'clickRight', 'clickTitle'],
 		props: {
+			dark: {
+				type: Boolean,
+				default: false
+			},
 			title: {
 				type: String,
 				default: ""
@@ -94,11 +105,11 @@
 			},
 			color: {
 				type: String,
-				default: "#000000"
+				default: ""
 			},
 			backgroundColor: {
 				type: String,
-				default: "#FFFFFF"
+				default: ""
 			},
 			statusBar: {
 				type: [Boolean, String],
@@ -111,13 +122,58 @@
 			border: {
 				type: [Boolean, String],
 				default: true
+			},
+			height: {
+				type: [Number, String],
+				default: 44
+			},
+			leftWidth: {
+				type: [Number, String],
+				default: 60
+			},
+			rightWidth: {
+				type: [Number, String],
+				default: 60
+			},
+		},
+		computed: {
+			themeBgColor() {
+				if (this.dark) {
+					// 默认值
+					if (this.backgroundColor) {
+						return this.backgroundColor
+					} else {
+						return this.dark ? '#333' : '#FFF'
+					}
+				}
+				return this.backgroundColor || '#FFF'
+			},
+			themeColor() {
+				if (this.dark) {
+					// 默认值
+					if (this.color) {
+						return this.color
+					} else {
+						return this.dark ? '#fff' : '#333'
+					}
+				}
+				return this.color || '#333'
+			},
+			navbarHeight() {
+				return getVal(this.height)
+			},
+			leftIconWidth() {
+				return getVal(this.leftWidth)
+			},
+			rightIconWidth() {
+				return getVal(this.rightWidth)
 			}
 		},
-        mounted() {
-          if(uni.report && this.title !== '') {
-              uni.report('title', this.title)
-          }
-        },
+		mounted() {
+			if (uni.report && this.title !== '') {
+				uni.report('title', this.title)
+			}
+		},
 		methods: {
 			onClickLeft() {
 				this.$emit("clickLeft");
@@ -135,6 +191,10 @@
 <style lang="scss" scoped>
 	$nav-height: 44px;
 
+	.uni-navbar {
+		// box-sizing: border-box;
+	}
+
 	.uni-nav-bar-text {
 		/* #ifdef APP-PLUS */
 		font-size: 34rpx;
@@ -150,37 +210,42 @@
 
 	.uni-navbar__content {
 		position: relative;
-		background-color: #fff;
-		overflow: hidden;
-		// width: 750rpx;
+		// background-color: #fff;
+		// box-sizing: border-box;
+		background-color: transparent;
 	}
 
 	.uni-navbar__content_view {
+		// box-sizing: border-box;
+	}
+
+	.uni-navbar-btn-text {
 		/* #ifndef APP-NVUE */
 		display: flex;
 		/* #endif */
+		flex-direction: column;
+		justify-content: flex-start;
 		align-items: center;
-		flex-direction: row;
-		// background-color: #FFFFFF;
+		line-height: 12px;
 	}
 
 	.uni-navbar__header {
-		padding: 0 10px;
 		/* #ifndef APP-NVUE */
 		display: flex;
 		/* #endif */
+		padding: 0 10px;
 		flex-direction: row;
 		height: $nav-height;
-		line-height: $nav-height;
 		font-size: 12px;
-		// background-color: #ffffff;
 	}
 
 	.uni-navbar__header-btns {
 		/* #ifndef APP-NVUE */
+		overflow: hidden;
 		display: flex;
 		/* #endif */
 		flex-wrap: nowrap;
+		flex-direction: row;
 		width: 120rpx;
 		// padding: 0 6px;
 		justify-content: center;
@@ -196,19 +261,27 @@
 		/* #endif */
 		width: 120rpx;
 		justify-content: flex-start;
+		align-items: center;
 	}
 
 	.uni-navbar__header-btns-right {
 		/* #ifndef APP-NVUE */
 		display: flex;
 		/* #endif */
-		width: 150rpx;
-		padding-right: 30rpx;
+		flex-direction: row;
+		// width: 150rpx;
+		// padding-right: 30rpx;
 		justify-content: flex-end;
+		align-items: center;
 	}
 
 	.uni-navbar__header-container {
+		/* #ifndef APP-NVUE */
+		display: flex;
+		/* #endif */
 		flex: 1;
+		padding: 0 10px;
+		overflow: hidden;
 	}
 
 	.uni-navbar__header-container-inner {
@@ -216,9 +289,12 @@
 		display: flex;
 		/* #endif */
 		flex: 1;
+		flex-direction: row;
 		align-items: center;
 		justify-content: center;
 		font-size: 12px;
+		overflow: hidden;
+		// box-sizing: border-box;
 	}
 
 
@@ -234,16 +310,14 @@
 		right: var(--window-right);
 		/* #endif */
 		/* #ifndef H5 */
-		left:0;
+		left: 0;
 		right: 0;
 		/* #endif */
 
 	}
 
 	.uni-navbar--shadow {
-		/* #ifndef APP-NVUE */
 		box-shadow: 0 1px 6px #ccc;
-		/* #endif */
 	}
 
 	.uni-navbar--border {
@@ -251,4 +325,19 @@
 		border-bottom-style: solid;
 		border-bottom-color: #eee;
 	}
+
+	.uni-ellipsis-1 {
+		overflow: hidden;
+		/* #ifndef APP-NVUE */
+		white-space: nowrap;
+		text-overflow: ellipsis;
+		/* #endif */
+		/* #ifdef APP-NVUE */
+		lines: 1;
+		text-overflow: ellipsis;
+		/* #endif */
+	}
+
+	// 暗主题配置
+	.uni-dark {}
 </style>