+++ uni_modules/uni-nav-bar/components/uni-nav-bar/uni-status-bar.vue
@@ -18,7 +18,7 @@
 	}
 </script>
 
-<style lang="scss" scoped>
+<style lang="scss" >
 	.uni-status-bar {
 		// width: 750rpx;
 		height: 20px;