+++ uni_modules/uni-popup/components/uni-popup-dialog/uni-popup-dialog.vue
@@ -15,7 +15,7 @@
 		</view>
 		<view class="uni-dialog-button-group">
 			<view class="uni-dialog-button" @click="closeDialog">
-				<text class="uni-dialog-button-text">{{cancelText}}</text>
+				<text class="uni-dialog-button-text">{{closeText}}</text>
 			</view>
 			<view class="uni-dialog-button uni-border-left" @click="onOk">
 				<text class="uni-dialog-button-text uni-button-color">{{okText}}</text>
@@ -31,7 +31,7 @@
 	initVueI18n
 	} from '@dcloudio/uni-i18n'
 	import messages from '../uni-popup/i18n/index.js'
-	const {	t	} = initVueI18n(messages)
+	const {	t } = initVueI18n(messages)
 	/**
 	 * PopUp 弹出层-对话框样式
 	 * @description 弹出层-对话框样式
@@ -84,6 +84,14 @@
 			beforeClose: {
 				type: Boolean,
 				default: false
+			},
+			cancelText:{
+				type: String,
+				default: ''
+			},
+			confirmText:{
+				type: String,
+				default: ''
 			}
 		},
 		data() {
@@ -95,10 +103,10 @@
 		},
 		computed: {
 			okText() {
-				return t("uni-popup.ok")
+				return this.confirmText || t("uni-popup.ok")
 			},
-			cancelText() {
-				return t("uni-popup.cancel")
+			closeText() {
+				return this.cancelText || t("uni-popup.cancel")
 			},
 			placeholderText() {
 				return this.placeholder || t("uni-popup.placeholder")
@@ -162,7 +170,7 @@
 	}
 </script>
 
-<style lang="scss" scoped>
+<style lang="scss" >
 	.uni-popup-dialog {
 		width: 300px;
 		border-radius: 11px;