+++ uni_modules/uni-upgrade-center-app/pages/upgrade-popup.vue
@@ -3,7 +3,7 @@
 		<view class="content botton-radius">
 			<view class="content-top">
 				<text class="content-top-text">{{title}}</text>
-				<image class="content-top" style="top: 0;" width="100%" height="100%" src="../static/bg_top.png">
+				<image class="content-top" style="top: 0;" width="100%" height="100%" src="../images/bg_top.png">
 				</image>
 			</view>
 			<view class="content-header"></view>
@@ -55,7 +55,7 @@
 				</view>
 			</view>
 
-			<image v-if="!is_mandatory" class="close-img" src="../static/app_update_close.png"
+			<image v-if="!is_mandatory" class="close-img" src="../images/app_update_close.png"
 				@click.stop="closeUpdate"></image>
 		</view>
 	</view>