+++ uni_modules/uni-list/changelog.md
@@ -1,3 +1,6 @@
+## 1.2.0（2021-11-23）
+- 优化 组件UI，并提供设计资源，详见:[https://uniapp.dcloud.io/component/uniui/resource](https://uniapp.dcloud.io/component/uniui/resource)
+- 文档迁移，详见:[https://uniapp.dcloud.io/component/uniui/uni-list](https://uniapp.dcloud.io/component/uniui/uni-list)
 ## 1.1.3（2021-08-30）
 - 修复 在vue3中to属性在发行应用的时候报错的bug
 ## 1.1.2（2021-07-30）